+++ packages/components/src/Shared/Ripples/index.tsx
@@ -8,9 +8,10 @@ import React, {
 } from "react";
 import styled from "@emotion/styled";
 import { motion } from "framer-motion";
-import type { HTMLMotionProps } from "framer-motion";
+import type { MotionProps } from "framer-motion";
 
-export interface RipplesProps extends HTMLMotionProps<"div"> {
+type Extendable = MotionProps & React.ComponentPropsWithoutRef<"div">;
+export interface RipplesProps extends Extendable {
   /** the animation duration of the ripple @default 600 */
   duration?: number;
   /** the color of the ripple, @default rgba(0, 0, 0, .3) */
@@ -21,6 +22,8 @@ export interface RipplesProps extends HTMLMotionProps<"div"> {
   children: React.ReactNode;
   /** the css border radius of the ripple, @default none */
   borderRadius?: CSSProperties["borderRadius"];
+  /** disable the ripple */
+  disabled?: boolean;
 }
 
 const boxStyle: CSSProperties = {
@@ -58,6 +61,7 @@ export const Ripples = memo(
     borderRadius = "none",
     onClick,
     children,
+    disabled,
     ...rest
   }: RipplesProps) => {
     const [rippleStyle, setRippleStyle] = useState<CSSProperties>({});
@@ -72,6 +76,7 @@ export const Ripples = memo(
     const onClickHandler = useCallback(
       (event: React.MouseEvent<HTMLDivElement>) => {
         event.stopPropagation();
+        if (disabled) return;
         // clear the timeout if exists
         if (timeoutId.current !== null) clearTimeout(timeoutId.current);
 
@@ -105,7 +110,7 @@ export const Ripples = memo(
 
         if (typeof onClick === "function") onClick(event);
       },
-      [color, duration, onClick]
+      [color, duration, disabled, onClick]
     );
 
     return (